Require NumPy
Makes NumPy a requirement for building the module. Sets the base NumPy
requirement at 1.8.2. This can always be adjusted later. Just need some
version of NumPy to be installed initially. As there are relatively
minimal requirements expected from NumPy for this library, starting at
the oldest version of NumPy available in conda-forge for Python 2.7 and
Python 3.4-3.6 seems like a good version of NumPy to start with. Can
bump this version up as needed.